pyproject.toml
[project]
name = "sldb"
version = "0.1.0"
description = "Structured Language Database - structured Markdown to Pydantic and back"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["markdown", "pydantic", "structured-data", "templating", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "pydantic>=2.0",
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
sldb = "sldb.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "build>=1.2",
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sldb = ["templates/*.md"]
"sldb.templates.example_bundle" = ["*.md", "*.yaml"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"
